#22 Change DecreasePercent from long? to decimal?
DecreasePercent should be decimal? to:
Support fractional percentages (e.g., 12.5%, 7.75%)
if needed in capital decrease operations
Follow financial domain best practices confirmed for
Iran's capital markets: use fixed-precision decimal
(DECIMAL/NUMERIC) with appropriate scale, not integers
or floating-point types
Maintain consistency with the codebase: ownershipPercentage
uses decimal, and the related IncreasePercent field
(in ExtraAssemblyIncreaseCapital.cs) uses double, both supporting fractional values

+++ src/Application/Fundamental.Application/Codals/Manufacturing/Queries/GetExtraAssemblyById/GetExtraAssemblyDetailItem.cs
@@ -91,8 +91,8 @@ public sealed record ExtraAssemblyIncreaseCapitalQueryDto(
 );
 
 public sealed record ExtraAssemblyDecreaseCapitalQueryDto(
-    long? CapitalDecreaseValue,
-    long? DecreasePercent,
+    decimal? CapitalDecreaseValue,
+    decimal? DecreasePercent,
     bool IsAccept,
     long? NewCapital,
     long? NewShareCount,
